+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes
 * SimulatorControl now also creates the tvOS simulator
 * Initial Carthage support. Carthage is not bootstraped, so make sure that Carthage is installed if your project uses it.
 * Embedded provisioning file now honors the team-id when expanding the identifiers
+* Updated that the test result is not taken from the xcodebuild output, but from the TestSummaries.plist, because there are causes where the xcodebuild output stopps.
 
 Note: There is no 0.14.1 version. (Reason is a typo)